requirements.txt
aiogram
google-genai
aiohttp
